a71-common: overlay: Enable UDFPS framework dimming
This is currently only supported on a handful of ROMs, but this allows the framework to do the screen dimming for optical in-screen fingerprint sensors instead of the kernel (which our device doesn't support).

The calculations are based from the Lineage 17.1 FOD HALs used for Samsung Galaxy As[1] and may be off, but these can be tuned once we tested this.

Change-Id: I0a59dee74a68daaf9b5635e3aa1e62a1a34d33fd